refactor: presentation/map/cluster 패키지 생성 및 관련 파일 이동
