+++ android/Staccato_AN/app/src/main/java/com/on/staccato/di/module/ManagerModule.kt
@@ -4,7 +4,7 @@ import android.content.ClipboardManager
 import android.content.Context
 import android.content.Context.CLIPBOARD_SERVICE
 import com.on.staccato.presentation.common.ShareManager
-import com.on.staccato.presentation.map.ClusterDrawManager
+import com.on.staccato.presentation.map.cluster.ClusterDrawManager
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
