+++ android/Staccato_AN/app/src/main/java/com/on/staccato/presentation/map/cluster/StaccatoMarkerRender.kt
@@ -1,4 +1,4 @@
-package com.on.staccato.presentation.map
+package com.on.staccato.presentation.map.cluster
 
 import android.content.Context
 import com.google.android.gms.maps.GoogleMap
